sae/swarm/docker_images/insert/requirements.txt
aiohttp
asyncio-nats-client
nats.py
mysqlclient==2.1.1
